requirements.txt
requests~=2.28.2
botcity-framework-web~=0.8.2
selenium~=3.141.0